Add mutex thing to DNS builder; change name of mail suppressor
